refactor(server): move verifyIdentityToken to Identity.verify
Extract the standalone identity-token verifier out of Handler.auth into a
dedicated `Identity` module exported from accounts/server. Callers now use
`Identity.verify(...)` instead of `Handler.verifyIdentityToken(...)`; the impl
lives in internal/handlers/identity.ts and Handler.auth imports it internally.
Updates type tests and docs (new Identity.verify reference page).

Amp-Thread-ID: https://ampcode.com/threads/T-019ea89f-ae30-73df-925c-ca505c59997f
