+++ src/server/Handler.ts
@@ -4,7 +4,7 @@ import type { ExtractSchema, MergeSchemaPath, Schema } from 'hono/types'
 import type { UnionToIntersection } from '../internal/types.js'
 import * as RequestListener from './internal/requestListener.js'
 
-export { auth, verifyIdentityToken } from './internal/handlers/auth.js'
+export { auth } from './internal/handlers/auth.js'
 export { type SessionRequest } from './internal/handlers/session.js'
 export { codeAuth } from './internal/handlers/codeAuth.js'
 export { exchange } from './internal/handlers/exchange.js'
