+++ src/server/Identity.ts
@@ -1,0 +1,1 @@
+export { verify } from './internal/handlers/identity.js'
