+++ src/server/index.ts
@@ -2,4 +2,5 @@ export { hc } from 'hono/client'
 
 export * as CliAuth from './CliAuth.js'
 export * as Handler from './Handler.js'
+export * as Identity from './Identity.js'
 export * as Kv from './Kv.js'
